Lights are cleared on level load, fixing a bug of old lights staying on new levels

+++ Managers/DebugWindow.cs
@@ -21,7 +21,7 @@ public partial class DebugWindow : Form
             action();
     }
 
-    public void SetFrameTimes(IEnumerable<(string Name, double Time)> frameTimes)
+    public void SetFrameTimes(IEnumerable<(string Name, float Time)> frameTimes)
     {
         RunOnUI(() =>
         {
